+++ src/univariate_milp_relaxation.jl
@@ -1,4 +1,16 @@
 """
+    _check_consistency(formulation_info, num_vars) 
+Checks consistency between provided variables and partition sizes 
+"""
+function _check_consistency(formulation_info::FormulationInfo, num_vars)::Bool
+    var = formulation_info.variables
+    if haskey(var, :z)
+        (length(var[:z]) == num_vars) && (return true)
+    end
+    return false
+end
+
+"""
     _build_univariate_milp_relaxation!(m,x,y,function_data,pre_base_name)
 
 Return a MILPRelaxation object with constraint and RHS information of the MILP
@@ -9,7 +21,8 @@ function _build_univariate_milp_relaxation!(
     x::JuMP.VariableRef,
     y::JuMP.VariableRef,
     univariate_function_data::UnivariateFunctionData,
-    pre_base_name::AbstractString,
+    variable_pre_base_name::AbstractString,
+    reuse::FormulationInfo,
 )::FormulationInfo
     sec_vs, tan_vs = _collect_vertices(univariate_function_data)
     formulation_info = FormulationInfo()
@@ -17,13 +30,17 @@ function _build_univariate_milp_relaxation!(
     # create variables
     num_vars = length(univariate_function_data.partition) - 1
 
+    reuse_variables = reuse.variables
+
+    is_consistent = _check_consistency(reuse, num_vars)
+
     delta_1 =
         formulation_info.variables[:delta_1] = JuMP.@variable(
             m,
             [1:num_vars],
             lower_bound = 0.0,
             upper_bound = 1.0,
-            base_name = pre_base_name * "delta_1"
+            base_name = variable_pre_base_name * "delta_1"
         )
     delta_2 =
         formulation_info.variables[:delta_2] = JuMP.@variable(
@@ -31,18 +48,20 @@ function _build_univariate_milp_relaxation!(
             [1:num_vars],
             lower_bound = 0.0,
             upper_bound = 1.0,
-            base_name = pre_base_name * "delta_2"
+            base_name = variable_pre_base_name * "delta_2"
         )
     z =
-        formulation_info.variables[:z] = JuMP.@variable(
+        (is_consistent) ? reuse_variables[:bin] :
+        JuMP.@variable(
             m,
             [1:num_vars],
             binary = true,
-            base_name = pre_base_name * "z"
+            base_name = variable_pre_base_name * "_z"
         )
+    formulation_info.variables[:bin] = z
 
     # add x constraints
-    formulation_info.constraints[:x] = JuMP.@constraint(
+    JuMP.@constraint(
         m,
         x ==
         sec_vs[1][1] + sum(
@@ -52,7 +71,7 @@ function _build_univariate_milp_relaxation!(
     )
 
     # add y constraints
-    formulation_info.constraints[:y] = JuMP.@constraint(
+    JuMP.@constraint(
         m,
         y ==
         sec_vs[1][2] + sum(
@@ -62,14 +81,10 @@ function _build_univariate_milp_relaxation!(
     )
 
     # add first delta constraint
-    formulation_info.constraints[:first_delta] =
-        JuMP.@constraint(m, delta_1[1] + delta_2[1] <= 1)
-
+    JuMP.@constraint(m, delta_1[1] + delta_2[1] <= 1)
     # add linking constraints between delta_1, delta_2 and z
-    formulation_info.constraints[:below_z] =
-        JuMP.@constraint(m, [i = 2:num_vars], delta_1[i] + delta_2[i] <= z[i-1])
-    formulation_info.constraints[:above_z] =
-        JuMP.@constraint(m, [i = 2:num_vars], z[i-1] <= delta_2[i-1])
+    JuMP.@constraint(m, [i = 2:num_vars], delta_1[i] + delta_2[i] <= z[i-1])
+    JuMP.@constraint(m, [i = 2:num_vars], z[i-1] <= delta_2[i-1])
 
     return formulation_info
 end